Add messages.

+++ Source/KungFu/AbstractLevelKungFu.cs
@@ -2,12 +2,12 @@ namespace y1000.Source.KungFu;
 
 public abstract class AbstractLevelKungFu : ILevelKungFu
 {
-    protected AbstractLevelKungFu(float level)
+    protected AbstractLevelKungFu(int level)
     {
         Level = level;
     }
     
-    public float Level { get; }
+    public int Level { get; }
 
     public abstract string Name { get; }
 }
